fixing karma configuration
karma-jspm is being used so karma-babel-preprocessor is not necessary
anymore as karma-jspm handles it.

The karma babel preprocessor was creating a conflict with karma-jspm,
so I removed babel configuration in karma and uninstalled the npm
package.

Also, for some reason the connect function of the exported object in
socket-io mock was undefined, so I created the object with the function
inside of the definition and export the instance of that object, now it
seems to be working.